+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/ui/fragments/FirstFragment.kt
@@ -49,7 +49,7 @@ class FirstFragment : Fragment() {
             SdkRepository(appContext),
             NetworkTestingRepository(appContext),
             AppExitRepository(),
-            StressTestRepository(),
+            StressTestRepository(appContext),
         )
     }